Update Product.js
Change datatype to float.

+++ src/db/models/Product.js
@@ -24,7 +24,7 @@ const Product = sequelize.define("products", {
     }
   },
   price: {
-    type: DataTypes.DECIMAL(7,2),
+    type: DataTypes.FLOAT,
     allowNull: false,
   }
 });
